+++ svgwrite/extensions/__init__.py
@@ -1,4 +1,3 @@
 # Copyright (c) 2018 Manfred Moitzi
 # License: MIT License
 from .inkscape import Inkscape
-from .shapes import *